[model, test] refactor: rename schedulers/flow_matching.py -> flow_matching_loss.py
Removes the near-identical name pair veomni.schedulers.flow_match (inference-time FlowMatchScheduler) vs veomni.schedulers.flow_matching (training-path helpers). The new name flow_matching_loss reads as a distinct concept, so the two modules are unambiguous at import sites.

- git mv veomni/schedulers/flow_matching.py veomni/schedulers/flow_matching_loss.py
- Update the two import sites: veomni/models/transformers/hunyuan_image_3/configuration_hunyuan_image_3.py and tests/models/test_hunyuan_image_3.py (+ a subprocess-invoked import string in the same file).
- HI3 patchgen config's add_import + a code comment; regenerate .py + .diff so the generated modeling reflects the new module path.
- veomni/schedulers/__init__.py docstring: drop the "near-identically named" warning that this rename makes redundant.
